Refactor penalty group creation/retrieval into dedicated service
