+++ src/functions/createPenaltyGroup.js
@@ -1,21 +1,13 @@
 import { doc } from 'serverless-dynamodb-client';
-import PenaltyDocument from '../services/penaltyDocuments';
+import PenaltyGroup from '../services/penaltyGroups';
 
-const penaltyDocuments = new PenaltyDocument(
+const penaltyGroups = new PenaltyGroup(
 	doc,
 	process.env.DYNAMODB_PENALTY_DOC_TABLE,
 	process.env.DYNAMODB_PENALTY_GROUP_TABLE,
-	process.env.BUCKETNAME,
-	process.env.SNSTOPICARN,
-	process.env.SITERESOURCE,
-	process.env.PAYMENTURL,
-	process.env.TOKEN_SERVICE_ARN,
-	process.env.DAYS_TO_HOLD || 3,
 );
 
 export default (event, context, callback) => {
-
 	const data = JSON.parse(event.body);
-	// id body document
-	penaltyDocuments.createPenaltyGroup(data, callback);
+	penaltyGroups.createPenaltyGroup(data, callback);
 };
