+++ src/functions/getPenaltyGroup.js
@@ -1,20 +1,12 @@
 import { doc } from 'serverless-dynamodb-client';
-import PenaltyDocument from '../services/penaltyDocuments';
+import PenaltyGroup from '../services/penaltyGroups';
 
-const penaltyDocuments = new PenaltyDocument(
+const penaltyDocuments = new PenaltyGroup(
 	doc,
 	process.env.DYNAMODB_PENALTY_DOC_TABLE,
 	process.env.DYNAMODB_PENALTY_GROUP_TABLE,
-	process.env.BUCKETNAME,
-	process.env.SNSTOPICARN,
-	process.env.SITERESOURCE,
-	process.env.PAYMENTURL,
-	process.env.TOKEN_SERVICE_ARN,
-	process.env.DAYS_TO_HOLD || 3,
 );
 
 export default (event, context, callback) => {
-
-	// id body document
 	penaltyDocuments.getPenaltyGroup(event.pathParameters.id, callback);
 };
